Item Capsules

See merge request KartKrew/Kart!440

+++ src/d_player.h
@@ -425,7 +425,7 @@ typedef struct player_s
 	UINT8 ringdelay;	// (0 to 3) - 3 tic delay between every ring usage
 	UINT16 ringboost;	// Ring boost timer
 	UINT8 sparkleanim;	// (0 to 19) - Angle offset for ring sparkle animation
-	UINT8 superring;	// Spawn rings on top of you every tic!
+	UINT16 superring;	// Spawn rings on top of you every tic!
 
 	UINT8 curshield;	// see kartshields_t
 	UINT8 bubblecool;	// Bubble Shield use cooldown
